refactor: simplify economy
